viewNext(), scrolling tagbar & ctrl-mod-left/right
viewNext focuses the tagset arg->i places to the right of the current
one. This is used to implement scrolling on the tag bar, and using
ctrl-mod-left/right to shift the viewed tags. This resembles workspace
behavior, but allows for multiple selected tags, e.g. if tags 1 and 3
were selected and viewNext was called with arg->i=1, tags 2 and 4 would
be selected instead. I have only thus far tested viewNext on +/-1 (via
the two shortcuts mentioned), but I believe it should work on any i from
-31 to 31 inclusive for use in more advanced workflows.